requirements.txt
attrs==23.1.0
blinker==1.6.2
boxsdk==3.7.3
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.5
cryptography==41.0.2
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
urllib3==1.26.16
Werkzeug==2.3.6
